+++ NEWS.md
@@ -1,3 +1,7 @@
+# Unreleased version
+
+- New feature: the screenshot image can now be saved on the server using `server_dir` parameter. Also added `download` parameter to specify whether to download the image to the app's user. (#3)
+
 # shinyscreenshot 0.1.0 (2020-10-13)
 
 Initial release